fix issue with images having relative url.
Issue : We load the content of pages by using ajax and we push the new content in the dom before to change the url. Some pages like the home page have image defined by a relative url. As the url is not the good one when we put the content in the DOM, the images are displayed as broken.

To reproduce the issue just click on an article (article -> overview) then click on the home link of the top menu. The images are not displayed correctly. Refresh the browser, the images appears.

Change-Id: I72a2997d9e6400b97b5e11da05dfd282d17882ff